Split project to 2 different projects
