+++ System/RobotControlAdapter/robotcontroladapter.cpp
